+++ lib/src/test/kotlin/com/maltaisn/recurpicker/RecurrenceTest.kt
@@ -63,33 +63,31 @@ internal class RecurrenceTest {
 
     @Test
     fun `should create string representation in debug builds`() {
-        if (BuildConfig.DEBUG) {
-            assertEquals("Recurrence{ Does not repeat }",
-                Recurrence.DOES_NOT_REPEAT.toString())
-            assertEquals("Recurrence{ Every 5 days }",
-                Recurrence(Period.DAILY) { frequency = 5 }.toString())
-            assertEquals("Recurrence{ Every year; until Dec 30, 2010 }",
-                Recurrence(Period.YEARLY) { endDate = dateFor("2010-12-30") }.toString())
-            assertEquals("Recurrence{ Every year; for 12 events }",
-                Recurrence(Period.YEARLY) { endCount = 12 }.toString())
-            assertEquals("Recurrence{ Every week on Sun, Wed }",
-                Recurrence(Period.WEEKLY) { setDaysOfWeek(Recurrence.SUNDAY, Recurrence.WEDNESDAY) }.toString())
+        assertEquals("Recurrence{ Does not repeat }",
+            Recurrence.DOES_NOT_REPEAT.toString())
+        assertEquals("Recurrence{ Every 5 days }",
+            Recurrence(Period.DAILY) { frequency = 5 }.toString())
+        assertEquals("Recurrence{ Every year; until Dec 30, 2010 }",
+            Recurrence(Period.YEARLY) { endDate = dateFor("2010-12-30") }.toString())
+        assertEquals("Recurrence{ Every year; for 12 events }",
+            Recurrence(Period.YEARLY) { endCount = 12 }.toString())
+        assertEquals("Recurrence{ Every week on Sun, Wed }",
+            Recurrence(Period.WEEKLY) { setDaysOfWeek(Recurrence.SUNDAY, Recurrence.WEDNESDAY) }.toString())
 
-            assertEquals("Recurrence{ Every month (on Monday of the third week) }",
-                Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.MONDAY, 3) }.toString())
-            assertEquals("Recurrence{ Every month (on Saturday of the last week) }",
-                Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.SATURDAY, -1) }.toString())
-            assertEquals("Recurrence{ Every month (on Wednesday of the fourth to last week) }",
-                Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.WEDNESDAY, -4) }.toString())
-            assertEquals("Recurrence{ Every month (on the same day each month) }",
-                Recurrence(Period.MONTHLY).toString())
-            assertEquals("Recurrence{ Every month (on the last day of the month) }",
-                Recurrence(Period.MONTHLY) { dayInMonth = -1 }.toString())
-            assertEquals("Recurrence{ Every month (on 16 days before the end of the month) }",
-                Recurrence(Period.MONTHLY) { dayInMonth = -16 }.toString())
-            assertEquals("Recurrence{ Every month (on the 25 of each month) }",
-                Recurrence(Period.MONTHLY) { dayInMonth = 25 }.toString())
-        }
+        assertEquals("Recurrence{ Every month (on Monday of the third week) }",
+            Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.MONDAY, 3) }.toString())
+        assertEquals("Recurrence{ Every month (on Saturday of the last week) }",
+            Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.SATURDAY, -1) }.toString())
+        assertEquals("Recurrence{ Every month (on Wednesday of the fourth to last week) }",
+            Recurrence(Period.MONTHLY) { setDayOfWeekInMonth(Recurrence.WEDNESDAY, -4) }.toString())
+        assertEquals("Recurrence{ Every month (on the same day each month) }",
+            Recurrence(Period.MONTHLY).toString())
+        assertEquals("Recurrence{ Every month (on the last day of the month) }",
+            Recurrence(Period.MONTHLY) { dayInMonth = -1 }.toString())
+        assertEquals("Recurrence{ Every month (on 16 days before the end of the month) }",
+            Recurrence(Period.MONTHLY) { dayInMonth = -16 }.toString())
+        assertEquals("Recurrence{ Every month (on the 25 of each month) }",
+            Recurrence(Period.MONTHLY) { dayInMonth = 25 }.toString())
     }
 
     @Test